requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
icalendar==5.0.12
beautifulsoup4==4.12.3
